change to a .ts file
